+++ cw_evm/lib/evm_erc20_balance.dart
@@ -4,7 +4,7 @@ import 'package:cw_core/balance.dart';
 
 class EVMChainERC20Balance extends Balance {
   EVMChainERC20Balance(this.balance, {this.exponent = 18})
-      : super(balance.toInt(), balance.toInt());
+      : super(balance, balance);
 
   final BigInt balance;
   final int exponent;
